Move get_recipe_list_page
